add the eval function and compareTo, best-first search

+++ 6_SlidingGame/src/pkg6_slidinggame/Configuration.java
@@ -21,6 +21,6 @@ public interface Configuration extends Comparable<Configuration>{
     public default List<Configuration> pathFromRoot(){
         return new ArrayList<>(); // there is no path because this is root
     }
-    public abstract double eval();
+    public abstract int eval();
     
 }
